% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{actionButtonStatus}
\alias{actionButtonStatus}
\title{Copied from shiny::actionButton, but provide support for status}
\usage{
actionButtonStatus(
  inputId,
  label,
  icon = NULL,
  width = NULL,
  status = "default",
  ...
)
}
\arguments{
\item{status}{Button color: "primary", "danger", "info", "success", "warning", "secondary", "dark" or "light"}
}
\description{
Copied from shiny::actionButton, but provide support for status
}
