% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_filter.R
\name{as.data.frame.DatasetFilter}
\alias{as.data.frame.DatasetFilter}
\title{Convert a DataFilter to data frame}
\usage{
\method{as.data.frame}{DatasetFilter}(x, ...)
}
\description{
Convert a DataFilter to data frame
}
