% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{columnLabels}
\alias{columnLabels}
\title{Gets the labels from a SAS data set (read in by haven::read_sas)
Returns the column names if there are no labels}
\usage{
columnLabels(tb, removeId = NULL)
}
\arguments{
\item{tb}{the tibble read in by haven::read_sas}
}
\description{
Gets the labels from a SAS data set (read in by haven::read_sas)
Returns the column names if there are no labels
}
