% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{columnLabelsInvert}
\alias{columnLabelsInvert}
\title{Gets the labels from a SAS data set in format for selectInput
list with key as label and value as the variable name}
\usage{
columnLabelsInvert(tb, removeId = NULL)
}
\arguments{
\item{tb}{the tibble read in by haven::read_sas}

\item{removeId}{the name of a column to remove.}
}
\description{
Gets the labels from a SAS data set in format for selectInput
list with key as label and value as the variable name
}
