% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{commonColumns}
\alias{commonColumns}
\title{Returns the common columns of a list of data frame (or other objects that has column names)}
\usage{
commonColumns(datalist)
}
\arguments{
\item{datalist}{list (or similar) of data frame (or similar) with column names}
}
\value{
A list with two keys: `common` that contains the common names as well as
}
\description{
Returns the common columns of a list of data frame (or other objects that has column names)
}
