% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createCommonTable}
\alias{createCommonTable}
\title{Create a 'subject level' table from a list of data sets
Assume that the data sets have some common variables
Extract these variables into a new data set}
\usage{
createCommonTable(datalist)
}
\arguments{
\item{datalist}{a list of data sets}
}
\description{
Create a 'subject level' table from a list of data sets
Assume that the data sets have some common variables
Extract these variables into a new data set
}
