% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_filter.R
\name{DatasetFilter}
\alias{DatasetFilter}
\title{Create object for data set filter}
\usage{
DatasetFilter(
  dataset_name,
  variable = NULL,
  filterType = c("categorical", "date", "numeric"),
  catValues = NULL,
  minVal = NULL,
  maxVal = NULL,
  includeMissing = FALSE,
  onlyMissing = FALSE
)
}
\arguments{
\item{dataset_name}{data set name  to filter on}

\item{variable}{name of variable from `dataset_name` that we are filtering on}

\item{filterType}{One of "categorical", "date", or "numeric"}

\item{catValues}{If categorical is selected for \code{filterType} should provide a vector of choices to select. A subject will match if \code{variable} is in \code{catValues}}

\item{minVal}{the minimum value to match if \code{filterType} is "date" or "numeric". Selected values greater than or equal to \code{minVal}}

\item{maxVal}{the maximum value to match if \code{filterType} is "date" or "numeric". Selected values less than or equal to \code{maxVal}}

\item{includeMissing}{Boolean. Should we include missing values? Defaults to \code{FALSE}.}

\item{onlyMissing}{Boolean. Should we only include subjects with missing values from \code{variable}? Defaults to \code{FALSE}. Need to provide the value filter. If "categorical" type is given, then catValues must be provided. If `date` or `numeric` is given, one or both of minVal or maxVal should be provided.}
}
\description{
Create object for data set filter
}
