% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{datatableControls}
\alias{datatableControls}
\title{Creates a data table with controls in a column
Warning: The created table has `escape = FALSE` by default since we are generating HTML code}
\usage{
datatableControls(df, id, ns = NS(""), ...)
}
\arguments{
\item{df}{The data frame we want to add the control column to}

\item{id}{id of the created columns}

\item{ns}{Namespace function}

\item{...}{The options to be passed into datatable.}
}
\description{
Creates a data table with controls in a column
Warning: The created table has `escape = FALSE` by default since we are generating HTML code
}
