% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractRowNumber}
\alias{extractRowNumber}
\title{Extract the row number from a id in a datatable when row number is the last digit in the id}
\usage{
extractRowNumber(idstr)
}
\description{
Extract the row number from a id in a datatable when row number is the last digit in the id
}
\examples{
\dontrun{
extractRowNumber("population-population_card-controls_delete_button_2")
}
}
