% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte_tables.R
\name{formatCI}
\alias{formatCI}
\title{Format confidence interval}
\usage{
formatCI(val, lower, upper, digits = 1)
}
\arguments{
\item{val}{The point estimate for the confidence inteval}

\item{lower}{The lower bound of the confidence interval}

\item{upper}{The upper bound of the confidence interval}

\item{digits}{the number of digits after the decimal place. Default 1}
}
\value{
string "val [lower, upper]"
}
\description{
Format confidence interval
}
