% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_filter.R
\name{getFilterVarNames}
\alias{getFilterVarNames}
\title{Gets a string of the variable names in the population filter}
\usage{
getFilterVarNames(popfilter)
}
\arguments{
\item{popfilter}{the PopulationFilter}
}
\description{
Gets a string of the variable names in the population filter
}
