% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hasClass}
\alias{hasClass}
\title{Check if a given css class is in the class string}
\usage{
hasClass(css_class, class_string)
}
\arguments{
\item{css_class}{The class we are looking for}

\item{class_string}{The class string that we want to find \code{css_class} in}
}
\value{
Boolean indicating if \code{css_class} is in \code{class_string}
}
\description{
Check if a given css class is in the class string
}
