% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.date}
\alias{is.date}
\title{Check if a value is a date}
\usage{
is.date(x)
}
\arguments{
\item{x}{object}
}
\value{
boolean indicating if x is a Date or not.
}
\description{
From: https://stackoverflow.com/a/37062951/1351718
}
