% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte_tables.R
\name{kmEventSummary}
\alias{kmEventSummary}
\title{Create the table summary for the events based on RATIFY's kmQ function}
\usage{
kmEventSummary(fit)
}
\arguments{
\item{fit}{output from \code{survival::survfit}}
}
\value{
data frame of the event summary. Includes the total number (N), the number of events (No. Events), and number censored (No. Censored) in each strata
}
\description{
Create the table summary for the events based on RATIFY's kmQ function
}
\examples{
library(survival)
fit <- survfit(Surv(time, status) ~ sex, data=survival::lung)
subpat::kmEventSummary(fit)
}
\seealso{
\url{https://gitlabce.statwb.eu.novartis.net/MAFE3/RATIFY/blob/ce87136a47eae7ec2d9e52df1d9879b00a05cc13/functions/31_km_functions_fm_ratify.R#L14}
}
