% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte_tables.R
\name{kmEventTables}
\alias{kmEventTables}
\title{Creates a event table from a survival model.
Return a table with cumulative number of  events, the number at risk, and the survival probability with confidence interval.}
\usage{
kmEventTables(
  fit,
  times = round(seq(min(fit$time), ceiling(max(fit$time)), length.out = 10))
)
}
\arguments{
\item{fit}{output from \code{survival::survfit}}

\item{times}{Optional vector of times to show in the table. By default will include equally spaced times from the minimum event time to the maximum event time with 10 time points.}
}
\value{
kable output of the KM event table from survfit \code{fit} at time points \code{times}
}
\description{
Creates a event table from a survival model.
Return a table with cumulative number of  events, the number at risk, and the survival probability with confidence interval.
}
\examples{
library(survival)
fit <- survfit(Surv(time, status) ~ sex, data=survival::lung)
subpat::kmEventTables(fit)
}
