% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte_tables.R
\name{kmQuartiles}
\alias{kmQuartiles}
\title{Create the table summary for the quartiles (Q1, Median, Q3), from RATIFY's kmQ function}
\usage{
kmQuartiles(fit)
}
\arguments{
\item{fit}{output from \code{survival::survfit}}
}
\value{
data frame of the event summary.
}
\description{
Create the table summary for the quartiles (Q1, Median, Q3), from RATIFY's kmQ function
}
\examples{
library(survival)
fit <- survfit(Surv(time, status) ~ sex, data=survival::lung)
subpat::kmQuartiles(fit)
}
\seealso{
\url{https://gitlabce.statwb.eu.novartis.net/MAFE3/RATIFY/blob/ce87136a47eae7ec2d9e52df1d9879b00a05cc13/functions/31_km_functions_fm_ratify.R#L9}
}
