% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labelToColumn}
\alias{labelToColumn}
\title{Gets the column from a given label}
\usage{
labelToColumn(dataset, lab)
}
\arguments{
\item{dataset}{the dataset (tibble, data frame)}

\item{lab}{the label we are looking for in the column names}
}
\value{
the column in the data set that matches \code{lab}
#' @seealso \code{\link{labelToName}} which return the column \emph{name}, rather than the column itself.
}
\description{
Gets the column from a given label
}
