% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labelToName}
\alias{labelToName}
\title{Gets the given variable name from a label}
\usage{
labelToName(dataset, lab)
}
\arguments{
\item{dataset}{the dataset (tibble, data frame)}

\item{lab}{the label we are looking for in the column names}
}
\value{
the column name in the data set that match \code{lab}
}
\description{
Gets the given variable name from a label
}
\seealso{
\code{\link{labelToColumn}} which return the actual column, rather than the column name.
}
