% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numericColNames}
\alias{numericColNames}
\title{Returns numeric column names}
\usage{
numericColNames(d, showLabels = TRUE)
}
\arguments{
\item{d}{the data set}

\item{showLabels}{should the labels be returned in addition to the values?
If TRUE, names(list) = labels(d)}
}
\description{
Returns numeric column names
}
