% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partition}
\alias{partition}
\title{Partitions a numerical vector into factors
Ensures that all values have a category (NA will still be NA)}
\usage{
partition(x, breaks, ...)
}
\description{
Partitions a numerical vector into factors
Ensures that all values have a category (NA will still be NA)
}
