% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{patientCountBadge}
\alias{patientCountBadge}
\title{Creates the patient count badge}
\usage{
patientCountBadge(p, total, name = NULL, as.string = FALSE)
}
\arguments{
\item{p}{the selected number of patients}

\item{total}{the total number of subjects}

\item{as.string}{if TRUE returns the html as a character}
}
\description{
Creates the patient count badge
}
