% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{population_control_column}
\alias{population_control_column}
\title{Creates the control column html buttons prefixed with a given id}
\usage{
population_control_column(ns = NS(NULL), id, len)
}
\arguments{
\item{ns}{the namespace (defaults to no namespace)}

\item{id}{the id (without namespace)}

\item{len}{the number of control elements to create}
}
\description{
When a delete button is pressed, `deletePopulationClicked` will change
When the edit button is pressed, `editPopulationClicked` will change
}
\details{
The checkbox will have id ns(id)_checkbox_index
The delete buttons will have id ns(id)_delete_button_index
The edit buttons will have id ns(id)_edit_button_index
}
