% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_filter.R
\name{PopulationFilter}
\alias{PopulationFilter}
\title{Constructor for "PopulationFilter" class. A list of filters with more specific functionality}
\usage{
PopulationFilter(
  dataset_names,
  name = NULL,
  initFilters = list(),
  createDate = date(),
  editDate = date()
)
}
\arguments{
\item{name}{the name of the population filter}

\item{initFilters}{a list of DatasetFilters}

\item{createDate}{the data this filter was created}

\item{datasets}{the named list of data frames we want to filter.}

\item{editData}{the date this filter was last edited}
}
\value{
PopulationFilter object
}
\description{
Constructor for "PopulationFilter" class. A list of filters with more specific functionality
}
