% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_filter.R
\name{query.DatasetFilter}
\alias{query.DatasetFilter}
\title{Query the dataset filter and get back the data set with filtered rows}
\usage{
\method{query}{DatasetFilter}(x, dataset, ...)
}
\arguments{
\item{x}{the DatasetFilter}

\item{dataset}{the dataset (data frame, tibble, etc.) that we want to filter}
}
\description{
Query the dataset filter and get back the data set with filtered rows
}
