% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_filter.R
\name{query.PopulationFilter}
\alias{query.PopulationFilter}
\title{Query the population filter}
\usage{
\method{query}{PopulationFilter}(popfilter, datasets, idkey = "USUBJID", commonTable = NULL)
}
\arguments{
\item{popfilter}{The PopulationFilter object}

\item{datasets}{a list of data sets to query}

\item{idkey}{the id key common to all data sets}
}
\description{
Query the population filter
}
