% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_filter.R
\name{queryIndividual.PopulationFilter}
\alias{queryIndividual.PopulationFilter}
\title{Gets all ids that match the current filters in the given filter collection
This applies all the filters for a given dataset only to that dataset.
Use \code{query.PopulationFilter} make sure filters apply cross-dataset}
\usage{
queryIndividual.PopulationFilter(
  popfilter,
  datasets,
  idkey = "USUBJID",
  commonTable = NULL
)
}
\arguments{
\item{idkey}{id key name for the filter}

\item{commonTable}{a string of the table name in the filter collection that contains all metadata about the subjects (E.g. ADSL)}

\item{PopulationFilter}{PopulationFilter object}
}
\value{
ids that match the current filter in `PopulationFilter`
}
\description{
Gets all ids that match the current filters in the given filter collection
This applies all the filters for a given dataset only to that dataset.
Use \code{query.PopulationFilter} make sure filters apply cross-dataset
}
