% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_filter.R
\name{querySubjects.DatasetFilter}
\alias{querySubjects.DatasetFilter}
\title{Get the unique patients that match the given filter}
\usage{
\method{querySubjects}{DatasetFilter}(x, dataset, idkey = "USUBJID", ...)
}
\arguments{
\item{x}{a DatasetFilter}
}
\description{
Get the unique patients that match the given filter
}
