% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_filter.R
\name{querySubjects.PopulationFilter}
\alias{querySubjects.PopulationFilter}
\title{Query the subjects from a population filter}
\usage{
\method{querySubjects}{PopulationFilter}(
  popfilter,
  datasets,
  idkey = "USUBJID",
  commonTable = NULL,
  mergeResults = TRUE
)
}
\arguments{
\item{popfilter}{the PopulationFilter object}

\item{datasets}{the datasets to filter on}

\item{idkey}{the subject identification key in the}

\item{mergeResults}{Boolean, should the results from each data set be merged? If TRUE, returns a single vector of subject ids
If FALSE, will return a list with keys of the data set names and a vector of ids for each data set}
}
\description{
Query the subjects from a population filter
}
