% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpat.R
\docType{package}
\name{subpat-package}
\alias{subpat}
\alias{subpat-package}
\title{subpat: Subpopulation and subgroups shiny modules}
\description{
subpat provides a collection of shiny modules to create subgroup and subpopulations as well as create ad-hoc table listings. These modules are featured in the two applications, the Patient Listing Generator (PLG) and the TTE analysis app.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://opensource.nibr.com/subpat}
  \item Report bugs at \url{https://github.com/Novartis/subpat/issues}
}

}
\author{
\strong{Maintainer}: Stefan Eng \email{stefaneng13@gmail.com} (\href{https://orcid.org/0000-0002-5245-6507}{ORCID})

Authors:
\itemize{
  \item Xiao Ni \email{xiao.ni@novartis.com}
  \item Mustapha Larbaoui \email{mustapha.larbaoui@novartis.com}
}

Other contributors:
\itemize{
  \item Novartis [copyright holder, funder]
}

}
\keyword{internal}
