% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{subpatExamples}
\alias{subpatExamples}
\title{Runs the various module examples}
\usage{
subpatExamples(i)
}
\arguments{
\item{i}{the index or name of the example}
}
\description{
Example applications:
Call with index subpatExamples(2) or name subpatExamples('06_tte_simple')
   01_add-filter-example
   02_edit-population-example
   03_subpopulation-example
   04_table-listing-example
   05_variable-selection-example
   06_tte_simple
}
