% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_filter.R
\name{toPandoc}
\alias{toPandoc}
\title{Returns a population filter as pandoc}
\usage{
toPandoc(popfilter)
}
\arguments{
\item{popfilter}{PopulationFilter}
}
\description{
Returns a population filter as pandoc
}
