% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_filter.R
\name{toString.DatasetFilter}
\alias{toString.DatasetFilter}
\title{toString implemention for DatasetFilter class}
\usage{
\method{toString}{DatasetFilter}(x, label = FALSE, ...)
}
\arguments{
\item{label}{boolean indicating whether or not we want to use the SAS label or variable name. Defaults to the short variable name.}
}
\description{
toString implemention for DatasetFilter class
}
