% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{totalSubjects}
\alias{totalSubjects}
\title{Get the total number of subjects across a list of data sets.
Takes the union of all}
\usage{
totalSubjects(datasets, idkey = "USUBJID")
}
\arguments{
\item{datasets}{a list of data sets (data frames, tibbles)}

\item{idkey}{the idkey for the subjects. Default to USUBJID}
}
\value{
a list of subjects
}
\description{
Get the total number of subjects across a list of data sets.
Takes the union of all
}
