% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{validateIcon}
\alias{validateIcon}
\title{Copied from shiny input-action.R
This is not exposed in shiny, but need it to reimplement actionButton}
\usage{
validateIcon(icon)
}
\description{
Copied from shiny input-action.R
This is not exposed in shiny, but need it to reimplement actionButton
}
